[graphite] Add non-aa bounds filling renderer
Initially this is only used for drawPaint() since that always fills
its clip and doesn't actually have any of it's AA.

This has a few benefits over the previous local-rect mapping case:
1. Inverse mapping the clip bounds to find a local rect that would
   cover the clip bounds post projection can lead to extreme geometry.
   It's also not always possible if the CTM has perspective.
2. This adds no per-pixel overhead for coverage, unlike the AA rect
   renderer so should be more efficient, particularly for large fills.
3. drawPaint() is unlikely to be in a situation where we'd be breaking
   up batches anyways.
4. Since it doesn't have coverage it will not trigger dual source
   blending or other dst read requirements when using kSrc blend modes
   so should hopefully reduce power/bandwidth when filling a surface
   initially in a way that didn't hit the clear load op path.

This CL also fixes perspective correct interpolation for the inverse
fill case by adjusting the homogenous coordinate of sk_Position to
match the perspective plane. This allows it to handle inverse or
regular fills without having to change the varying to a vec3 and
do a divide in the fragment shader.

Bug: b/333729316
Change-Id: I31517417fdc545848d82c5dff53ad600746db892
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/872283
Reviewed-by: Jim Van Verth <jvanverth@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>