[Graphite][SkSL] Use std430 for storage buffers in Vulkan
Before SPIR-V would always default to std140 for anything that is not push constants. So we can still target SPIR-V 1.0, I added a struct to track if a StorageClassUniform is for a storage buffer so that the appropriate layout can be selected for pointers. This makes SSBOs always use the std430 layout to define struct member alignment and array stride.

Since this would change the default layout of SSBOs in SPIR-V, the corresponding VulkanCap for ssbo layout must also be changed in this CL.

Bug: b/243078879
Change-Id: Iec65ffe39ffe5bf3809bd4afb75799acd7f25456
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/870820
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>