Mention Java requirement in Android build instructions
This came up in I3ae70362348c7685bbc1afb12c86724ccf634f56, when I
needed to update my Java in order to use the patch, and I noticed that
we make no reference to requiring Java. Document that you need Java and
provide a link to download it.

While we're at it, make it clear that you need a *recent* NDK, as older
versions may stop working over time. (As of this writing, we've found
that 21.3.6528147 does *not* work, and 23.0.7599858 does, though the
latter will likely become out of date in the future.)

Change-Id: I91ea6cd550b509fb29b6047bdcf9881bf15c9ce7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/875237
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>
Commit-Queue: Leon Scroggins <scroggo@google.com>
Reviewed-by: Nolan Scobie <nscobie@google.com>