Swap radius scale factors in 270/90 rotation for SkRRect::transform
The x and y scale factors are applied to the dst (already swapped)
radii, so they must also be swapped.

This was not caught in the rrect transform unit tests for two reasons:
1. The existing flip+rotate tests were assuming left-to-right matrix
multiplication (or confusing pre vs. post operations on the matrix),
so the corner ordering was wrong. Without fixing the expectations for
those cases, they would fail with the swapped radii, but looking at
the actual matrix they were validating an incorrect dst rrect.

   However, the equivalant classes were all identified correctly, so
   even after working through the right-to-left matrix expectations,
   those tests are valid.

2. There were no unit tests for non-uniform scale factors, so I added
unit tests for flips, rotates, and non-uniform scales.

Bug: b/352692514
Change-Id: Ib74b8970925a418a506ad41b44cbfdca7cf134ad
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/876626
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>