Remove graphite-specific Mtl type aliases
These had been needed because these headers were used in C++ code.
Since the previous refactoring CLs to remove the SK_METAL defines
from the code, the separation is no longer needed and we can just
use the native types.

Note that I do add an #if guard to MtlGraphiteTypes.h because otherwise
the header will try to be used by fiddle.cpp which depends on a
a generated skia.h which just includes *all* of our public headers.

Of note, this does increase the stack size of the TextureInfo by 16
bytes because of the change from unsigned ints to their actual enums.

See also https://crrev.com/c/5707508 for an upstream change required
to land this.

Bug: b/294543706
Change-Id: I3bd2a25a71d209be14f590d2bb2bb13e4e63011a
Canary-Chromium-CL: 5707508
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/876660
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>