Add SkSL type alias for 'atomic_uint.'
`atomic_uint` is a GLSL ES3 reserved word, so SkSL needs to be
aware of it and treat it as a typename instead of as a freely usable
identifier. We now treat it as an alias for `atomicUint` (even though
its semantics don't match GL).

Bug: b/343194653
Change-Id: Ida1dda38c4a2715afe3442be7034a0d5a39d8a52
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/877337
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>