[graphite] Support writing substructs in UniformManager
Adds beginStruct(baseAlignment) and endStruct(), where it's planned
that the base alignment is taken from the cached alignment stored on
each shader snippet entry.

Updates UniformExpectationsValidator to know whether or not the
uniforms should be a substruct to confirm the offsets still match,
and that the struct is begun/ended appropriately.

At this point, it should be possible for built-in snippets to start
opting into using structs for their uniforms.

Bug: b/349208089
Change-Id: I419d6d18e9ffb17680f4ce3c667911db954e9c22
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/871378
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>