[graphite] Automate begin/endStruct() calls in KeyHelpers
Updates the VALIDATE_UNIFORMS to include both validation and
calling beginStruct() and endStruct() when it goes out of scope if
the snippet has a struct name declared.

This does now add a lookup to the shader code dictionary for every
block, so I opted to make the getEntry() that takes a built-in ID
in the header so that it can be easily inlined.

Bug: b/349208089
Change-Id: I1a35de74e133e329d3338ebb639c3b3f731d5049
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/876077
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>