[graphite] Snap non-AA stroke rects
When drawing a non-AA axis-aligned stroke rect, the local rect is
outset by the stroke radius and then rounded to pixels. From there
the ceiling of the device-space stroke width is calculated so that
the target inner edge is also pixel aligned.

Reconstructing the local coordinates and local stroke width undo
the scale factors on the rounded/ceil'ed values. For non-uniform
scales, this does mean that the two modified device stroke radii
cannot be represented as a single local stroke radius.

After playing around with animating scales and translations,
prioritizing the larger axis leads to the best results. Since the
exterior of the stroke is still pixel aligned, the resulting
emulated non-AA is effective.

Bug: b/280054774
Change-Id: I0f38d51b8e207e364e75b6184d52903240bff91f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/877416
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>