[SkSL][SPIR-V] Add function specialization for samplers
This removes samplers as parameters in the final SPIR-V assembly and instead creates different specializations of function that have samplers in their parameter list, inlining the sampler uniform directly.

This functionality is made generic so the same function to find specializations can be extended to child runtime effects and unsized arrays in the future.

Bug: b/349840463
Change-Id: I7ad3cbff51e1b7ab083606842cfd6b1f3c20b438
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/874997
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>