[skif] Track if padding was added to a layer; inset on restore() only if needed
If the outset can't be represented in 32-bit ints, then we can't
assume during the restore that the transparent padding exists.

This explicitly tracks whether or not the outset was added to the layer
bounds and only performs insetForSaveLayer() if it prepped the
layer's device correctly.

Bug: b/353980389
Change-Id: I56e53203a96cb10b568b55b0056d7758b5a3a6c9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/879256
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>