Use more stable keying for function-call specialization info.
Previously, we used a FunctionCall* as the key into our call
specialization map. This worked, but was slightly fragile
because backends can occasionally rewrite expressions (e.g. using
`clone`) and this would not be reflected in the map.

Now, we use a FunctionDeclaration* and a Position as the map key.
These are both relatively stable by comparison--we don't rewrite
entire functions, and the Position of an IRNode won't arbitrarily
move after it's been parsed.

Bug: b/349840463
Bug: b/328051548
Change-Id: I8b4e5865f4e95e7a3e1c5b271c51e3ec65549294
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/879276
Reviewed-by: Nathan Sanchez <nathanasanchez@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>