[graphite] Add AtlasProvider::freeGpuResources()
The TextAtlasManager had a freeAll() method that releases its textures,
but it was not being called in Recorder::freeGpuResources().

The RasterPathAtlas had no way to release its textures. This adds
resetting capabilities to PathAtlas and expands the existing
AtlasProvider::clearTexturePool() to handle clearing all GPU resources
that any of the atlases might reference.

ComputePathAtlases are somewhat transient and owned by DrawContexts,
and are reset each time the DrawContext is flushed so I don't think
they need to have their plot tracking or shape caching reset. Their
textures are un'refed with clearing the texture pool.

Bug: b/337980338
Change-Id: I878a3e2e61c55d64f45904c7df5fcfdbc842ebc6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/879656
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>