[graphite] Read aggregate SSBO once in main()
This modifies the snippet uniform lookups to plumb a struct parameter
through the generated functions to access values for the current draw.
Previously, the ssbo index had to be a mutable global variable that was
assigned in main() so that any function could then use it to read from
the SSBO as needed.

Now the SSBO is read once using the flat varying directly, and the
read struct is passed to the generated functions as needed. This had
a slight improvement in instruction count and occupancy on Metal, but
also produces more readable code when having to debug.

Bug: b/349207406
Change-Id: I71aeeae4d5f419afadb0874ffcd78c358c997811
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/876078
Reviewed-by: John Stiles <johnstiles@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>