Keep track of the current ModuleType as it is being compiled.
Previously, we knew whether or not a module was being compiled
(via fIsBuiltinCode), but not _which_ module. Since we have a
ModuleType enum now, it makes sense to actually keep it in the
ProgramConfig instead of just a boolean.

Change-Id: Iedf31a9f3003a40367eb1eca1070b7dc2e782fb1
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/879936
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>