[SkSL][SPIR-V] Reuse synthesized samplers per specialization.
When a synthesized sampler/texture pair was created for the function parameter variable which is specialized, it would override the mapping of any previous pair which caused a use after free later in the program. This is due to the parameter variable pointer which is the key matching even though it's a different specialization of the function being emitted.

To resolve this, I reused the synthesized texture/sampler parameter if it already exists. This works because the parameter is only needed in the variable map for the scope of `writeFunctionInstantiation`.

Bug: oss-fuzz:70523
Change-Id: I4aec0d41b55252a968051c03af1ae9e1b7d5ca99
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/879640
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>