[skif] Use src blending when possible
This should allow intermediate passes to operate with HW blending
disabled.

Non-AA draws with the default blend mode now switch to kSrc if it's
not being drawn to the final device (which keeps the default as
src-over to blend with prior content). For intermediate passes, the
prior content is already cleared to transparent black so non-aa
src blending and src-over blending are equivalent.

To avoid dual source blending or even more expensive blending
methods, the default is kept as src-over when there is AA.

Bug: b/347031058
Change-Id: Ic3d613c8422b0b8fec870772b8bba3cd5c6b3f0a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/880656
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>