Fix fuzzer-discovered issue with struct redefinition.
Previously, `Type::matches` was implemented in terms of the name
of a type. However, a malformed program can define two structs
with the same name, but different contents, and declare variables
associated with each struct. Those two variables should be seen
as incompatible/impossible to coerce, but because their names
matched, `Type::matches` returned true and `coerceExpression`
succeeded. This led to an assertion soon afterwards.

We now require types to be an exact match. Arrays and aliases are
treated with special care, however. (In particular, array types
are created on demand, and it is currently possible to create more
than one ArrayType for a given type/count pair.)

Bug: oss-fuzz:70589
Change-Id: I612137b4afb8369d57b9fb808c94071c10ad763d
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/880501
Commit-Queue: Brian Osman <brianosman@google.com>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>