Track function calls via a stable ID instead of via pointer.
Using a FunctionCall* as a map key is fragile, since a clone of
the function call's IRNode should be functionally identical to
the original.

We now assign each FunctionCall node a "stable ID" consisting of
the Position of the original function call in the source, and the
module that the function call originated in. This stable ID will
persist even across runs of the inliner. (The inliner rewrites
Positions on IRNodes so that inlining errors are flagged at the
call site.)

Change-Id: If85db881354343bd7338075004f43176320992b6
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/880896
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Nathan Sanchez <nathanasanchez@google.com>