[SkSL][Metal] Add unsized array parameter support in Metal
As part of enabling unsized array parameters in SkSL, this CL removes some of the checks in the IR frontend and implements the proper type for  pointers in Metal.

To ensure this is the correct syntax and works in practice, I locally tested the array parameter in a graphite shader module and ran a gm that uses that array and things worked as expected.

Bug: b/343510513
Change-Id: I5168fa772f292e7d0acb548b1e72dfc3467010f4
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/879259
Reviewed-by: John Stiles <johnstiles@google.com>