[SkSL] Change specialization mapping to expression
Since we plan to add unsized array parameter support to SPIR-V, we need to be able to accept `FieldAccess` expressions as well. Since a `FieldAccess` to an unsized array doesn't have a variable to the array but instead, a variable to the base type and an index into the array member. The specialization mapping should be changed to an expression to support `VariableReference` and `FieldAccess`.

Change-Id: I56419bc112833831e8c279f6c5482e67b2a5ae5f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/881256
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>