[graphite] Add comments to public API re Graphite-specific deprecations
This CL also updates the VkProtectedContext_AsyncReadFromProtectedSurface* unit tests to actually test something. Prior to this CL those tests were creating an 8x8 surface and then trying to readback to a 10x10 result. The higher level error checking would reject that before even getting to the Protected-ness checks. Additionally, Graphite has deprecated the SkSurface::asyncRescaleAndReadPixelsYUV420 call so that would've resulted in a trivial result for Graphite. For both Ganesh and Graphite the unit tests should now be, correctly, failing the async readback due to the Protected-ness of the source SkSurface.

Bug: b/353638204
Change-Id: If25f76febd3ccfeba2b150210083d929dfff0077
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/881956
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Robert Phillips <robertphillips@google.com>