[graphite] Fill interior of large [r]rects with non-aa renderer
This greatly helps performance of the BouncingImages motionmark test
as well as other GMs like `many_circles` that feature lots of
overlapping but opaque draws.

The inner fill gets sorted early and prevents the majority of overdraw
while the AA exterior completes the drawing with anti-aliasing, but
because it shares the same Z value as the inner fill and uses a
GREATER test, the majority of pixels with an expensive fragment
shader are skipped.

The fragment shader can be expensive just because it's the
AnalyticRRectRenderStep (particularly if it's in a curved corner /
circle case) or because the paint is pretty expensive. Ganesh
with non-AA draws often avoids overdraw on tiling GPUs because of
FPK because it can disable blending. This depth-testing approach
gets almost the same amount of overdraw reduction but works on
hardware that doesn't support FPK and doesn't lose the AA quality.

Frame times for the bouncing image slide went from 38ms to 9ms
on my Mac M1 mini (times measured before Jim's improvements to the
YUV shader and slide). The many_circles GM went from a frame time
of 70ms to 10ms.

Bug: b/280054774
Bug: b/306635070
Change-Id: Ida6f1ca450e5671057ad6490b0b4b1dea4b97204
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/877420
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
(cherry picked from commit 9f775551cb825607d95db55cc377803666b8e17e)
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/884018