Allow `writeFunctionCallArgument` to add multiple arguments.
Previously, this function would return a single SpvId which the
caller was responsible for adding to the argument-list. Except,
occasionally we needed to add _two_ arguments, and in this case
the function had a special bonus out-parameter to return the
second argument as well.

More recently, some (specialization-related) paths now return NA,
and we needed to check for NA explicitly in order to avoid adding
it to the argument list.

The code now uses a simpler approach--the caller passes in the
argument list (which is always represented as `TArray<SpvId>`)
and `writeFunctionCallArgument` just adds in the number of
parameters that are needed. This allows for any number of
arguments, instead of adding in special cases.

Change-Id: Ide15631519d569e5078eafa40d30eb198bb800af
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/881996
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>
Reviewed-by: Nathan Sanchez <nathanasanchez@google.com>