fix: include package.json in exports field
This patch is a small change to the exports field. 

By adding the package.json to the exports field, it is possible to import the package.json in JavaScript code. 

Without it, the exports field disallows importing the package.json and throws an error. 

Importing and reading the package.json from code is important to be able to access version information, metadata, etc. for bundlers, libraries, projects, etc.

Change-Id: Ib71a0de78e965336da72e338d13c5c31cdcc8be0
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/878858
Reviewed-by: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Jorge Betancourt <jmbetancourt@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>