[Graphite] Reorder local matrix matrix inversion
Previously we were inverting the gradient matrix so when it is concatenated to the local matrix and inverted again, it would be the original gradient matrix since it would've been inverted twice. From a chromium gradient test, there is a case where the gradient matrix is not well defined given large floats due to floating point math. The degenerate matrix would not be able to be inverted causing an assert.

This change inverts the local matrix before multiplying by the gradient matrix, saving an inversion when drawing gradients.

Bug: b/355186743
Change-Id: I809aeaae88e5a53c5f0d1f2190e3d587733669e2
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/881758
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: Nathan Sanchez <nathanasanchez@google.com>