[graphite] Add forcePathAA to subrun control
Renames SDFTControl to SubRunControl. Adds a field for forcing
anti-aliasing when glyphs are drawn as paths.

Graphite sets this flag so that glyphs that happen to be rectangular
paths that happen to come from fonts with the alias edge mode are not
drawn with pixel snapping, since the rest of the glyphs in the sub
run are highly unlikely to also be rectangular.

Bug: b/354058834
Change-Id: I19de5f6c651062818d09047d7e58f2cd6b410c91
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/883516
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Jim Van Verth <jvanverth@google.com>