Change vello's #ifdef guards
I don't know if this is the ideal fix, but it's at least using defines
set by the vello dependency and not Skia itself, so the flow
of defines is Bazel-compliant.

This also makes the GN and Bazel build of Vello set those defines
explicitly so our code can use said defines.

Bug: b/294543706
Change-Id: Ice200d24591bcfd92bef1a73d66b8c911dae00f7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/883877
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>