[skif] Fix identity scale and no-effects detection
When an axis has a rescale() factor of 1.0, but was odd, the centering
dst rect would end with 0.5 on each edge and then be rounded out. This
would cause the next step to shift the source image by half a pixel on
the axis that should be a no-op. Visually this error was often hidden
because the other axis was applying a larger scale amount, or both
axes would be 1.0 and skip the rescale steps entirely.

But there were also cases where we thought we had to do colorspace
conversion when either the source or the dst color spaces were null.
This updates the check to better match the conventions of legacy
color space behavior and detect more of those cases as true no-ops.

In early patch sets of https://skia-review.googlesource.com/c/skia/+/866556
the CS check mismatch was triggering test failures on the
serialize-8888 config because we went from a no-op case to a CS
mismatch case, skipping the no-op. Then, because of the 1/2 pixel
shift in the identity blur, the resulting image was slightly different
than the original.

Change-Id: I015aa03ed966b8575468b38d8d86282c96f26e84
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/884159
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>