[skif] Implement raster SkBlurEngine
This is not as large of a CL as it looks. The static C++ code inside
SkBlurImageFilter.cpp was copy-pasted into SkBlurEngine.cpp and then
its main entry point was wrapped in the SkBlurEngine::Algorithm
signature. The main edits to that code was switching back to using
pure SkIRects from LayerSpace<SkIRects>.

In terms of behavior, when opting into the blur engine approach
(default on Skia tree, staged behind a flag in chrome, android doesn't
use CPU blurs so not staged), the large SIMD blurs will be rescaled
to fit the GaussPass window size. This means that with the new engine,
the TentPass is never utilized.

In animating blurs I found the pop between TentPass and GaussPass
sigmas to be very noticeable. Rescaling and then doing a smaller
GaussPass is also faster.

If the initial blur radius is small enough, or if the color type
is not 8888, then the existing shader blur algorithm is selected
instead. One change here is that technically now the SIMD algorithm
will be chosen for both BGRA8 and RGBA8 code whereas previously it
was just for N32.

Once this lands and the old path is removed, it will be theoretically
possible for the N32 restriction for raster image filters to be
removed.  The TentPass implementation can also be deleted then.
In follow ups, the A8 Gaussian blur implementation used for CPU
mask filters can be migrated to the blur engine as a specialization
for that "color" type.

This CL also removes the last defines for GPU backend from the
src/effects directory.

Change-Id: I9be29ba75e0fbb2c9112634ae0a4125531cff00f
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/866556
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Robert Phillips <robertphillips@google.com>