Improve error reporting when a prototype disagrees with its function.
Previously, if we encountered a prototype of a function with
different modifier flags than its definition, we would just emit
"duplicate definition of 'foo'", which was misleading. It seemed
like an overzealous error report, because it's safe to prototype
a function after its definition. However, in actuality, the
prototype _mismatched_ the definition, and that should trigger an
error.

We now emit a clearer error message: "functions '...' and '...'
differ only in modifiers".

We also now test that it is safe to prototype a function after
its definition (assuming the prototype truly matches).

Bug: b/355329784
Change-Id: Id9642f77d50878c2648dca40aab4bb910de92bb8
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/883177
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: James Godfrey-Kittle <jamesgk@google.com>
Commit-Queue: James Godfrey-Kittle <jamesgk@google.com>