[graphite] Address Vulkan Protected-mode Buffer Validation bug
We were, initially, creating un-Protected buffers which the VVLs were flagging that with the error:

   vkCmdCopyBuffer(): dstBuffer is an unprotected buffer, but command
   buffer is protected. The Vulkan spec states: If commandBuffer is a
   protected command buffer and protectedNoFault is not supported,
   dstBuffer must not be an unprotected buffer.

Additionally, the prior code was making the Protected buffers mappable.

This issue was reported for Graphite but I have attempted to keep the Ganesh and Graphite code in sync.

Bug: b/353638204

Change-Id: I19be404f63ea69b3d87474d7592d53fe63ca5d2b
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/884176
Commit-Queue: Robert Phillips <robertphillips@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>