[graphite] Add size to BindBufferInfo
This removes ClearBufferInfo, BindUniformBufferInfo, and BufferView.
All of these amounted to the existing BindBufferInfo and a size.

BindBufferInfo now stores the offset and the size as uint32_ts, so
the overall size is the same on a 64-bit machine.

Updates the BufferManagers to store the size of each suballocation in
the returned BindBufferInfo. Updates DrawWriter to add validation as
best as possible that the vertex/instance buffers have the right
amount of data in them.

Bug: b/296282257
Change-Id: I6f7244f586c63e94ba943ff3a086ce3764d350ff
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/882838
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>