Decouple Graphite SkSL code from Ganesh code.
The #ifdef approach won't work with the Bazel approach and we want
to be able to build both Ganesh and Graphite into the same binary
without conflicting symbols.

This makes the Graphite backend register the Graphite-specific
SkSL modules before they are needed. We think the first time a
graphite::Context is created is the cleanest way to do this.

Change-Id: I37a821db61d8f66b12bf669d0c8bd5064eafb576
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/884096
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>
Reviewed-by: John Stiles <johnstiles@google.com>