[graphite] Mark deferred Device immutable in snap()
Due to how references were held explicitly in the Recorder, this
Device wouldn't be cleaned up in the initial call to
flushTrackedDevices(). Subsequent calls to snap() would call
flushTrackedDevices() a second time, and in its order of operations,
would revisit the prior deferred Device that had its fully lazy
proxy moved with the Recording (possibly cross threads).

Only after this race-prone visit, would it detect that the Device was
uniquely held and clean it up.

Normal devices would be marked immutable in these situations because
an SkSurface would be being deleted. This adds matching logic so the
deferred Device is de-registered inside the snap() that moves its
lazy target to the Recording.

Change-Id: I9eda8e478d2175858cac3c050af03ebd813fafc7
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/886916
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Greg Daniel <egdaniel@google.com>