[graphite] Evict atlases on failed Recorder::snap()
Besides leading to a glyph or path mask not being in the texture on
subsequent Recordings, if there are no new atlas uploads added on the
second Recording, nothing will actually trigger instantiation of the
atlas texture proxies.

The instantiation happens as part of the upload preparation so may
not occur on a prior failure. Subsequently DrawPass can assert that
the atlas texture being sampled has never been instantiated.

Bug: b/339980098
Change-Id: I7e2f3989e16db8fa3f1e50d9e19a4eb474db3249
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/886956
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>