[graphite] Use EdgeAAQuad with all/none flags in drawSpecial
The image filtering code uses non-AA and kSrc blending in places to
fill pixels efficiently when it's determined that coverage isn't
needed. Graphite's drawSpecial() implementation was always using
the AARRectRenderStep unless it happened to be exactly pixel aligned
(in which case it would simplify to just the CoverBoundsRenderStep).

By switching to an EdgeAAQuad, drawSpecial() is now consistent with
drawImageRect when the paint is anti-aliased and consistent with
a non-AA per-edge AA quad coming into drawEdgeAAImageSet. It also
makes its coverage behavior match that of Ganesh and raster so we
won't inadvertantly produce a dst-read or dual-source blending
pipeline when drawing a "non-AA" special image that didn't quite
land on pixel bounds (e.g. final rotated transform with an image
that had transparent padding that can be sampled).

Change-Id: I9d10ff47a0344f80ac784daeac156c21bced1166
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/887879
Reviewed-by: Robert Phillips <robertphillips@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>