TArray: Remove some redundant ASAN unpoisoning
With ASAN enabled, we end up with frequent calls to poison and unpoison,
even in common operations like push_back. This CL doesn't remove all of
them, but addresses the very common case of calling push_back
repeatedly, where we would unpoison twice in a row (at the start of
push_back, and again in changeSize). Hopefully that mitigates the issue
enough that the oss-fuzz infrastructure can finish in < 60 s.

On my machine, this change reduces the run time of the linked fuzzer
input from 42 seconds to 25 seconds.

Bug: oss-fuzz:71112
Change-Id: Ia2bedf2e04cdcb7d337120552f55d133dd0adbd9
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/888817
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>