Combine GR_TEST_UTILS and GRAPHITE_TEST_UTILS into GPU_TEST_UTILS
We have currently decided to allow #if defined(GPU_TEST_UTILS) to
guard test-only code, which means we need test-only versions of
skia_cc_libraries, e.g. ganesh_TEST_UTIL, or so called testing-shadow
clones of libraries. To reduce the complexity of this, we are
combining the two defines into one so we only need 1 shadow universe
and not 3 (Ganesh only, Graphite only, Ganesh+Graphite).

This adds in a PRESUBMIT.py to catch uses of the old #define from
slipping in.

The changes are nearly all mechanical, except for a few deduplicated
instances and one place that didn't have a defined() around it.

Bug: b/294543706
Change-Id: Ifdf914ac750944bfe0e77ca78c373715e79e2873
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/888976
Reviewed-by: Greg Daniel <egdaniel@google.com>
Commit-Queue: Kaylee Lubick <kjlubick@google.com>