Fix Bazel build of viewer after GPU_TEST_UTILS landed
GPU_TEST_UTILS being in SkImage_Base.h was problematic (it was
previously ok because it used to be GRAPHITE_TEST_UTILS and that
is not defined by Bazel [yet]) because we were sometimes but not
always compiling the file with the define set.

The result was viewer would assert because SkImage_Base->isRaster()
was not returning the correct result.

In general, we don't want to set GPU_TEST_UTILS except in the
places where we have a mirrored set of Bazel skia_cc_library rules
(e.g. src/gpu/ganesh:ganesh_TEST_UTIL). I added a PRESUBMIT.py
to catch infractions and fixed a few places where we didn't want
it set (or didn't need to set it, in the case of GrTypesPriv.h).

Bug: b/256870320
Change-Id: I2948a486da6ec70dd09200739c9e60168070c755
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/889357
Commit-Queue: Kaylee Lubick <kjlubick@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>