Implement asAColorMode and asAColorMatrix for SkWorkingFormatColorFilter
This is slightly lossy - the client knows that the filter "is" one of
these two, but doesn't learn about the working space. For Android (the
primary user of this API), that's fine. They wrap all color filters to
operate in sRGB, so nothing is really lost.

Bug: 360020740
Change-Id: Ib412ef17d51c68b631e7524ef5236b358cfdaf4a
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/890318
Commit-Queue: Brian Osman <brianosman@google.com>
Reviewed-by: Florin Malita <fmalita@google.com>