Fix undefined SkData class build issue with GCC 13.2
With GCC 13.2 the compiler complains that SkData is undefined while
trying to process the sp_sp<SkData> template instantiation used as
return value for the SkPDFSubsetFont() function declaration. GCC 14.2
does not exhibit this issue.

The workaround is to #include the header for SkData in SkPDFSubsetFont.h
to ensure the SkData class is defined at that point.

This problem was caught by the WebKit EWS builders while attempting to
update the copy of Skia carried in the WebKit source tree:

  In file included from /app/webkit/Source/ThirdParty/skia/src/pdf/SkPDFSubsetFont.h:6,
                   from /app/webkit/Source/ThirdParty/skia/src/pdf/SkPDFSubsetFont.cpp:4:
  /app/webkit/Source/ThirdParty/skia/include/core/SkRefCnt.h: In instantiation of ‘void SkSafeUnref(T*) [with T = SkData]’:
  /app/webkit/Source/ThirdParty/skia/include/core/SkRefCnt.h:256:20:   required from ‘sk_sp<T>::~sk_sp() [with T = SkData]’
  /app/webkit/Source/ThirdParty/skia/src/pdf/SkPDFSubsetFont.cpp:115:12:   required from here
  /app/webkit/Source/ThirdParty/skia/include/core/SkRefCnt.h:151:14: error: invalid use of incomplete type ‘class SkData’

Complete build logs:

  https://ews-build.webkit.org/#/builders/5/builds/66127
  https://ews-build.webkit.org/#/builders/2/builds/67857

Change-Id: Ib1b178ad5add31b8557ca8998c8817babc51b7de
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/889296
Auto-Submit: Adrian Perez de Castro <aperez@igalia.com>
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Commit-Queue: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Brian Osman <brianosman@google.com>