[graphite] Update createGraphicsPipeline trace slice to _ALWAYS variant
The name is a slight misnomer, as this doesn't mean the cost of tracing
will always be paid. It only means that it bypasses our concept of a
"minimal set" of trace events in the Android framework (since Skia's
own categories aren't conceptually compatible with ATrace), so if a
tracing session is active and targeting this process, then this event
will be traced without needing any developer-focused system properties
to be toggled first.

Note: relies on If073bfd45762922f101540a20896d7a99f50a575 to ensure
trace arguments aren't evaluated if using ATrace but !ATRACE_ENABLED()

This aligns with how pipeline creation is traced in Ganesh.

Bug: b/337301999
Change-Id: I5597cfc3d41631a2caeda6a7ac1e04b8d5259335
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/891676
Reviewed-by: Michael Ludwig <michaelludwig@google.com>
Reviewed-by: Nicolette Prevost <nicolettep@google.com>
Commit-Queue: Nolan Scobie <nscobie@google.com>