Stop forcing debug symbol generation with skia_enable_optimize_size
This (setting dwarf4) was done for better compatibility with bloaty on
the CodeSize jobs. However, we now always specify dwarf4 (when adding
the debug_symbols config), so this is unnecessary. Previously, this
meant that skia_enable_optimize_size would have the side effect of
adding symbols (and making the resulting binaries look much bigger).

Change-Id: Ia4eeed8c5022ff90f3ee0d263f8407c4f0e8a999
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/892217
Reviewed-by: Ben Wagner <bungeman@google.com>
Commit-Queue: Brian Osman <brianosman@google.com>