[pdf] Avoid known image/profile channel mismatch
PDF requires that any ICC profile for an image be stored separately from
the image itself. Even if the image data contains an ICC profile it will
be ignored and only the one in the PDF will be used. This means that PDF
interpreters tend to apply the ICC profile themselves instead of using a
standard image decoder with profile support.

The issue is when the profile on the SkImage or even in JPEG data does
not actually apply to the image data. This is most commonly seen in
single channel images with three channel ICC profiles embedded. This
often happens with Chromium which currenly marks all SkImages as sRGB
even when they don't have three channels, and at least one JPEG file
found in the wild which had a single channel but a three channel ICC
profile embedded.

When the number of channels mismatches a variety of strange behaviors
are observed such as ignoring the profile, shading the image red,
showing a vague error the user and behaving strangely, or showing three
copies of the image. To avoid this, check to see if the number of
channels in the ICC profile can be determined. If it does not agree with
the number of channels in the image them omit the ICC profile. If we
cannot determine the number of channels in the ICC profile then allow
the ICC profile to pass through.

In addition to fixing the issues seen in bug reports, this fixes the
encode-srgb-jpg gm, which internally creates a one channel jpeg with an
sRGB color space. Prior to this change the PDF generated from this gm
acts very strangly in Reader due to the channel mismatch.

Bug: 41496311
Bug: 361116767
Change-Id: Ia5535900bbc77f87d1548fa35ed0bdcc7a057c04
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/892186
Reviewed-by: Brian Osman <brianosman@google.com>
Commit-Queue: Ben Wagner <bungeman@google.com>